Extract control classes from ControlPanel.cpp
