+++ src/core/gui/ControlPanel.cpp
@@ -25,7 +25,10 @@
 
 #include "ControlPanel.h"
 
+#include "Button.h"
+#include "Knob.h"
 #include "LayoutDescription.h"
+#include "Popup.h"
 #include "core/Configuration.h"
 #include "core/controls.h"
 #include "core/gettext.h"
@@ -33,283 +36,6 @@
 
 #include <utility>
 #include <vector>
-
-void modal_midi_learn(Param);
-
-class SkinnedParameterControl : public juce::Component, protected UpdateListener
-{
-public:
-	SkinnedParameterControl(Parameter &p, juce::Image image, const LayoutDescription::Resource &r)
-	: parameter(p)
-	, frame_(0)
-	, image_(std::move(image))
-	, width_(r.width)
-	, height_(r.height)
-	, frames_(r.frames)
-	{
-		setSize(width_, height_);
-		parameter.addUpdateListener(this);
-	}
-
-	~SkinnedParameterControl() override
-	{
-		parameter.removeUpdateListener(this);
-	}
-
-	Parameter &parameter;
-
-	bool isPlugin {false};
-
-	void showPopupMenu() const
-	{
-		if (isPlugin) {
-			return;
-		}
-		auto menu = juce::PopupMenu();
-		auto p = parameter.getId();
-		menu.addItem(gettext("Assign MIDI Controller..."), true, false, [p] {
-			modal_midi_learn(p);
-		});
-		bool ignored = Preset::shouldIgnoreParameter(p);
-		menu.addItem(gettext("Ignore Preset Value"), true, ignored, [ignored, p] {
-			Preset::setShouldIgnoreParameter(p, !ignored);
-			Configuration &config = Configuration::get();
-			config.ignored_parameters = Preset::getIgnoredParameterNames();
-			config.save();
-		});
-		menu.showMenuAsync(juce::PopupMenu::Options());
-	}
-
-protected:
-	void mouseDown(const juce::MouseEvent &event) override
-	{
-		if (event.mods.isPopupMenu()) {
-			showPopupMenu();
-		}
-	}
-
-	void mouseDoubleClick(const juce::MouseEvent &event) override
-	{
-		parameter.setValue(parameter.getDefault());
-	}
-
-	void paint(juce::Graphics &g) override
-	{
-		int x = 0, y = 0;
-		if (image_.getHeight() == height_) {
-			x = frame_ * width_;
-		} else {
-			y = frame_ * height_;
-		}
-		g.drawImage(image_, 0, 0, width_, height_, x, y, width_, height_, false);
-	}
-
-	void UpdateParameter(Param param, float controlValue) override
-	{
-		int frame = int(float(frames_ - 1) * parameter.getNormalisedValue());
-		if (frame_ == frame) {
-			return;
-		}
-		frame_ = frame;
-		repaintFromAnyThread();
-	}
-
-	void repaintFromAnyThread()
-	{
-		auto mm = juce::MessageManager::getInstanceWithoutCreating();
-		if (mm && mm->isThisTheMessageThread()) {
-			repaint();
-		} else {
-			juce::MessageManager::callAsync([this] {
-				repaint();
-			});
-		}
-	}
-
-private:
-	std::atomic_int frame_;
-	juce::Image image_;
-	int width_, height_;
-	int frames_;
-};
-
-class Knob : public SkinnedParameterControl
-{
-public:
-	Knob(Parameter &parameter, juce::Image image, const LayoutDescription::Resource &r)
-	: SkinnedParameterControl(parameter, std::move(image), r)
-	{}
-
-	class Label : public juce::Component
-	{
-	public:
-		Label(juce::Component *parent) : parent_(parent)
-		{
-			setAlwaysOnTop(true);
-			setOpaque(true);
-			setVisible(false);
-			parent->addChildComponent(this);
-		}
-
-		int yInset {0};
-
-		void show(juce::Component *control, juce::String text)
-		{
-			if (control_ && control_ != control) return;
-			control_ = control;
-
-			text_ = text;
-			if (text.isEmpty())
-			{
-				setVisible(false);
-				return;
-			}
-
-			auto ttb = getLookAndFeel().getTooltipBounds(text, control->getPosition(), parent_->getLocalBounds());
-			auto width = std::max(control->getWidth(), ttb.getWidth());
-			auto inset = (control->getWidth() - width) / 2;
-			setBounds(control->getX() + inset, control->getBottom() - yInset, width, ttb.getHeight());
-			setVisible(true);
-			repaint();
-		}
-
-		void hide()
-		{
-			setVisible(false);
-			control_ = nullptr;
-			text_ = "";
-		}
-
-	private:
-		void paint(juce::Graphics &graphics)
-		{
-			getLookAndFeel().drawTooltip(graphics, text_, getWidth(), getHeight());
-		}
-
-		juce::Component *parent_{nullptr};
-		juce::Component *control_{nullptr};
-		juce::String text_;
-	};
-
-	Label *label {nullptr};
-
-protected:
-	juce::String getLabelText()
-	{
-		char text[32] = "";
-		return parameter_get_display(parameter.getId(), parameter.getValue(), text, sizeof text) ? text : "";
-	}
-
-	void mouseEnter(const juce::MouseEvent &event) override
-	{
-		label->show(this, getLabelText());
-	}
-	
-	void mouseExit(const juce::MouseEvent &event) override
-	{
-		label->hide();
-	}
-
-	void mouseDown(const juce::MouseEvent &event) override
-	{
-		if (event.mods.isPopupMenu()) {
-			showPopupMenu();
-			return;
-		}
-		referenceVal_ = parameter.getNormalisedValue();
-		referenceY_ = event.y;
-		parameter.willChange();
-	}
-
-	void mouseDrag(const juce::MouseEvent &event) override
-	{
-		SkinnedParameterControl::mouseDrag(event);
-		if (event.mods.isPopupMenu()) {
-			return;
-		}
-		float sensitivity;
-		if (parameter.getStep() == 0.f) {
-			sensitivity = 300
-					* (event.mods.isCtrlDown() ? 4.f : 1.f)
-					* (event.mods.isShiftDown() ? 4.f : 1.f);
-		} else {
-			sensitivity = std::min(40 * float(parameter.getSteps()), 480.f);
-		}
-
-		auto offset = float(referenceY_ - event.y) / sensitivity;
-		auto newVal = referenceVal_ + offset;
-		if (newVal != referenceVal_) {
-			parameter.setNormalisedValue(referenceVal_ + offset);
-			label->show(this, getLabelText());
-			referenceVal_ = newVal;
-			referenceY_ = event.y;
-		}
-	}
-
-	void mouseWheelMove(const juce::MouseEvent &event,
-						const juce::MouseWheelDetails &wheel) override
-	{
-		if (event.mouseWasDraggedSinceMouseDown()) {
-			return;
-		}
-		auto delta = wheel.deltaY / 2.f
-				/ (event.mods.isCtrlDown() ? 4.f : 1.f)
-				/ (event.mods.isShiftDown() ? 4.f : 1.f);
-		parameter.willChange();
-		parameter.setNormalisedValue(parameter.getNormalisedValue() + delta);
-		label->show(this, getLabelText());
-	}
-
-private:
-	float referenceVal_{0.f};
-	int referenceY_{0};
-};
-
-class Button : public SkinnedParameterControl
-{
-public:
-	Button(Parameter &parameter, juce::Image image, const LayoutDescription::Resource &r)
-	: SkinnedParameterControl(parameter, std::move(image), r) {}
-
-protected:
-	void mouseDown(const juce::MouseEvent &event) override
-	{
-		if (event.mods.isPopupMenu()) {
-			showPopupMenu();
-			return;
-		}
-		parameter.willChange();
-		parameter.setNormalisedValue(parameter.getNormalisedValue() > 0.f ? 0.f : 1.f);
-	}
-};
-
-class Popup : public SkinnedParameterControl
-{
-public:
-	Popup(Parameter &parameter, juce::Image image, const LayoutDescription::Resource &r)
-	: SkinnedParameterControl(parameter, std::move(image), r) {}
-
-protected:
-
-	void mouseDown(const juce::MouseEvent &event) override
-	{
-		if (event.mods.isPopupMenu()) {
-			showPopupMenu();
-			return;
-		}
-		auto strings = parameter_get_value_strings(parameter.getId());
-		auto menu = juce::PopupMenu();
-		for (int i = 0; i <= parameter.getSteps(); i++) {
-			auto value = parameter.getMin() + parameter.getStep() * float(i);
-			auto isTicked = parameter.getValue() == value;
-			menu.addItem(strings[i], true, isTicked, [&parameter = parameter, value] {
-				parameter.willChange();
-				parameter.setValue(value);
-			});
-		}
-		menu.showMenuAsync(juce::PopupMenu::Options());
-	}
-};
 
 class Skin
 {
@@ -386,13 +112,12 @@ struct ControlPanel::Impl final
 			const auto &resource = control.resource;
 			auto image = skin.getImage(resource);
 
-			std::unique_ptr<SkinnedParameterControl> component;
+			std::unique_ptr<Control> component;
 			if (control.type == "button") {
 				component = std::make_unique<Button>(parameter, image, resource);
 			}
 			if (control.type == "knob") {
-				component = std::make_unique<Knob>(parameter, image, resource);
-				dynamic_cast<Knob *>(component.get())->label = &label_;
+				component = std::make_unique<Knob>(parameter, image, resource, &label_);
 			}
 			if (control.type == "popup") {
 				component = std::make_unique<Popup>(parameter, image, resource);
